requirements.txt
streamlit
pandas
plotly
requests
streamlit-option-menu
SQLAlchemy
psycopg2-binary
